requirements.txt
torch>=1.12.0
scikit-learn>=1.0.0
numpy>=1.20.0
pandas>=1.3.0
tqdm>=4.60.0
PyYAML>=6.0
tensorboard>=2.8.0
matplotlib>=3.5.0
seaborn>=0.11.0
thop>=0.1.1